Autonomous Car Neural Network Model
In this project, I made use of the udacity simulator to obtain the required data (Driving images), so that I could make use of it, to train my Convolutional Neural Network and hence was able to predict the path that had to be followed by the car in the udacity virtual environment. The CNN model was formed, based on the NVDIA research paper for autonomous driving, provided as an open source paper.